Reorganize query_and_update module
