+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/query.rs
@@ -1,6 +1,6 @@
 use cdk_framework::act::node::canister_method::{CanisterMethod, CanisterMethodType, QueryMethod};
 
-use super::BuildCanisterMethods;
+use super::shared::BuildCanisterMethods;
 use crate::TsAst;
 
 impl TsAst {
