+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/shared/build_canister_methods.rs
