+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/shared/mod.rs
@@ -1,0 +1,4 @@
+pub(super) mod body;
+mod build_canister_methods;
+
+pub use build_canister_methods::BuildCanisterMethods;
